Fixed the NVS issue when using esp-idf 4.0+
This caused incompatibility with some apps built with esp-idf 3.2 such as Go-Play.

The cause was that we were overwriting the app's nvs partition because esp 4.0+ defaults to a partition named "nvs" whereas 3.3 defaulted to the first nvs partition of any name (nvs_fw here).

Solution is to specify the NVS part name in the code.

All of our multi-firmware's partitions are now prefixed with mfw_ as well.