Added self-flashing code for easy installation/updating
The app can now be built as a normal .fw and, after installation via SD Card, it will flash itself to the appropriate location.

This is useful because the serial flashing process can be complicated for users (especially with the MRGC).

I'll add instructions to build the .fw as soon as I'm reasonably sure that it is safe.